requirements.txt
discord.py
pandas-ta
alpaca-py
pandas